zappaAlbumFinder Summary
This program allows a user to search for a Frank Zappa album by entering a song or year, or by searching a track list by entering an album.

There are only 2 albums at the moment, Apostrophe and Joe's Garage.

The album class contains a constructor with 3 parameters (title, year, track listing).

The driver class uses the album class to create album objects. The methods within the driver class compare user input to the constructor parameters to output what the user is searching for.